add tests and re-factor
