+++ .isort.cfg
@@ -1,2 +1,2 @@
 [settings]
-known_third_party = pandas,pendulum,pytest,pytz,requests,setuptools
+known_third_party = iris,pandas,pendulum,pytest,pytz,requests,setuptools,xarray
